requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
pillow==10.1.0
opencv-python==4.8.1.78
gunicorn==21.2.0